restructure cal.py
