+++ .pre-commit-config.yaml
@@ -7,7 +7,7 @@ repos:
   - repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.11.9
     hooks:
-      - id: ruff-check
+      - id: ruff
         args: [--config, "pyproject.toml", --fix]
       - id: ruff-format
         args: [--config, "pyproject.toml"]
