+++ src/icalendar/__init__.py
@@ -2,19 +2,17 @@ from icalendar.alarms import (
     Alarms,
     AlarmTime,
 )
-from icalendar.cal import (
-    Alarm,
+from icalendar.cal.alarm import Alarm
+from icalendar.cal.calendar import (
     Calendar,
-    Component,
-    ComponentFactory,
-    Event,
-    FreeBusy,
-    Journal,
-    Timezone,
-    TimezoneDaylight,
-    TimezoneStandard,
-    Todo,
 )
+from icalendar.cal.component import Component
+from icalendar.cal.component_factory import ComponentFactory
+from icalendar.cal.event import Event
+from icalendar.cal.free_busy import FreeBusy
+from icalendar.cal.journal import Journal
+from icalendar.cal.timezone import Timezone, TimezoneDaylight, TimezoneStandard
+from icalendar.cal.todo import Todo
 from icalendar.enums import CUTYPE, FBTYPE, PARTSTAT, RANGE, RELATED, RELTYPE, ROLE
 from icalendar.error import (
     ComponentEndMissing,
@@ -66,63 +64,63 @@ from icalendar.timezone import use_pytz, use_zoneinfo
 from .version import __version__, __version_tuple__, version, version_tuple
 
 __all__ = [
-    "Calendar",
-    "Event",
-    "Todo",
-    "Journal",
-    "Timezone",
-    "TimezoneStandard",
-    "TimezoneDaylight",
-    "FreeBusy",
-    "Alarm",
-    "ComponentFactory",
-    "vBinary",
-    "vBoolean",
-    "vCalAddress",
-    "vDatetime",
-    "vDate",
-    "vDDDLists",
-    "vDDDTypes",
-    "vDuration",
-    "vFloat",
-    "vInt",
-    "vPeriod",
-    "vWeekday",
-    "vFrequency",
-    "vRecur",
-    "vText",
-    "vTime",
-    "vUri",
-    "vGeo",
-    "vUTCOffset",
-    "Parameters",
-    "q_split",
-    "q_join",
-    "use_pytz",
-    "use_zoneinfo",
-    "__version__",
-    "version",
-    "__version_tuple__",
-    "version_tuple",
-    "TypesFactory",
-    "Component",
-    "vMonth",
-    "IncompleteComponent",
-    "InvalidCalendar",
-    "Alarms",
-    "AlarmTime",
-    "ComponentEndMissing",
-    "ComponentStartMissing",
-    "IncompleteAlarmInformation",
-    "LocalTimezoneMissing",
     "CUTYPE",
     "FBTYPE",
     "PARTSTAT",
     "RANGE",
-    "vSkip",
     "RELATED",
-    "vSkip",
     "RELTYPE",
     "ROLE",
-    "FeatureWillBeRemovedInFutureVersion"
+    "Alarm",
+    "AlarmTime",
+    "Alarms",
+    "Calendar",
+    "Component",
+    "ComponentEndMissing",
+    "ComponentFactory",
+    "ComponentStartMissing",
+    "Event",
+    "FeatureWillBeRemovedInFutureVersion",
+    "FreeBusy",
+    "IncompleteAlarmInformation",
+    "IncompleteComponent",
+    "InvalidCalendar",
+    "Journal",
+    "LocalTimezoneMissing",
+    "Parameters",
+    "Timezone",
+    "TimezoneDaylight",
+    "TimezoneStandard",
+    "Todo",
+    "TypesFactory",
+    "__version__",
+    "__version_tuple__",
+    "q_join",
+    "q_split",
+    "use_pytz",
+    "use_zoneinfo",
+    "vBinary",
+    "vBoolean",
+    "vCalAddress",
+    "vDDDLists",
+    "vDDDTypes",
+    "vDate",
+    "vDatetime",
+    "vDuration",
+    "vFloat",
+    "vFrequency",
+    "vGeo",
+    "vInt",
+    "vMonth",
+    "vPeriod",
+    "vRecur",
+    "vSkip",
+    "vSkip",
+    "vText",
+    "vTime",
+    "vUTCOffset",
+    "vUri",
+    "vWeekday",
+    "version",
+    "version_tuple",
 ]
