+++ src/icalendar/cal/journal.py
@@ -1,0 +1,128 @@
+""""""
+
+from __future__ import annotations
+
+from datetime import date, datetime, timedelta
+from typing import Optional
+
+from icalendar.attr import (
+    categories_property,
+    color_property,
+    create_single_property,
+    exdates_property,
+    rdates_property,
+    rrules_property,
+    sequence_property,
+    summary_property,
+    uid_property,
+)
+from icalendar.cal.component import Component
+from icalendar.error import IncompleteComponent
+
+
+class Journal(Component):
+    """A descriptive text at a certain time or associated with a component.
+
+    A "VJOURNAL" calendar component is a grouping of
+    component properties that represent one or more descriptive text
+    notes associated with a particular calendar date.  The "DTSTART"
+    property is used to specify the calendar date with which the
+    journal entry is associated.  Generally, it will have a DATE value
+    data type, but it can also be used to specify a DATE-TIME value
+    data type.  Examples of a journal entry include a daily record of
+    a legislative body or a journal entry of individual telephone
+    contacts for the day or an ordered list of accomplishments for the
+    day.
+    """
+
+    name = "VJOURNAL"
+
+    required = (
+        "UID",
+        "DTSTAMP",
+    )
+    singletons = (
+        "CLASS",
+        "COLOR",
+        "CREATED",
+        "DTSTART",
+        "DTSTAMP",
+        "LAST-MODIFIED",
+        "ORGANIZER",
+        "RECURRENCE-ID",
+        "SEQUENCE",
+        "STATUS",
+        "SUMMARY",
+        "UID",
+        "URL",
+    )
+    multiple = (
+        "ATTACH",
+        "ATTENDEE",
+        "CATEGORIES",
+        "COMMENT",
+        "CONTACT",
+        "EXDATE",
+        "RELATED",
+        "RDATE",
+        "RRULE",
+        "RSTATUS",
+        "DESCRIPTION",
+    )
+
+    DTSTART = create_single_property(
+        "DTSTART",
+        "dt",
+        (datetime, date),
+        date,
+        'The "DTSTART" property for a "VJOURNAL" that specifies the exact date at which the journal entry was made.',  # noqa: E501
+    )
+
+    @property
+    def start(self) -> date:
+        """The start of the Journal.
+
+        The "DTSTART"
+        property is used to specify the calendar date with which the
+        journal entry is associated.
+        """
+        start = self.DTSTART
+        if start is None:
+            raise IncompleteComponent("No DTSTART given.")
+        return start
+
+    @start.setter
+    def start(self, value: datetime | date) -> None:
+        """Set the start of the journal."""
+        self.DTSTART = value
+
+    end = start
+
+    @property
+    def duration(self) -> timedelta:
+        """The journal has no duration: timedelta(0)."""
+        return timedelta(0)
+
+    color = color_property
+    sequence = sequence_property
+    categories = categories_property
+    rdates = rdates_property
+    exdates = exdates_property
+    rrules = rrules_property
+    uid = uid_property
+
+    summary = summary_property
+    # description = description_property # This does not work. It should be a list.
+
+    @classmethod
+    def new(cls, /, summary: Optional[str] = None):
+        """Create a new journal entry with all required properties.
+
+        This creates a new Journal in accordance with :rfc:`5545`.
+        """
+        journal = cls()
+        journal.summary = summary
+        return journal
+
+
+__all__ = ["Journal"]
