+++ src/icalendar/tests/prop/test_property_values.py
@@ -2,7 +2,7 @@
 
 from datetime import datetime
 
-from icalendar import Event
+from icalendar.cal.event import Event
 
 
 def test_vDDDLists_timezone(tzp):
