+++ src/icalendar/tests/test_issue_336_dateutil_timezone.py
@@ -9,10 +9,10 @@ from datetime import datetime
 
 from dateutil import tz
 
-from icalendar import Event
+from icalendar.cal.event import Event
 from icalendar.timezone import tzid_from_tzinfo
 
-valid_names = ("America/Noronha",  "Brazil/DeNoronha")
+valid_names = ("America/Noronha", "Brazil/DeNoronha")
 
 
 def test_timezone_name_directly():
