+++ src/icalendar/tests/test_issue_798_property_parameters.py
@@ -1,4 +1,7 @@
-"""Test the property parameters."""
+"""Test the property parameters.
+
+See https://github.com/collective/icalendar/issues/798
+"""
 
 from datetime import datetime
 
@@ -20,7 +23,6 @@ from icalendar.timezone.tzp import TZP
 
 
 class Prop:
-
     params: Parameters
 
     def __init__(self, **parameters):
@@ -41,7 +43,7 @@ class Prop:
     )
 
 
-@pytest.fixture()
+@pytest.fixture
 def p():
     """Empty test property."""
     return Prop()
@@ -52,15 +54,18 @@ def test_set_altrep(p):
     assert p.params == {"ALTREP": "http://example.com"}
     assert p.ALTREP == "http://example.com"
 
+
 def test_altrep_must_be_quoted():
     """altrepparam = "ALTREP" "=" DQUOTE uri DQUOTE"""
     assert Prop(ALTREP="1234aA").to_ical() == 'ALTREP="1234aA"'
+
 
 def test_del_altrep(p):
     """Del when empty"""
     del p.ALTREP
     assert p.params == {}
     assert p.ALTREP is None
+
 
 def test_del_altrep_full(p):
     """Del when empty"""
@@ -86,23 +91,26 @@ def test_set_delegation_to_string(p):
     p.DELEGATED_FROM = "mailto:foo"
     assert p.DELEGATED_FROM == ("mailto:foo",)
 
+
 def test_set_delegation_to_tuple(p):
-    p.DELEGATED_TO = ("mailto:foo","mailto:bar")
-    assert p.DELEGATED_TO == ("mailto:foo","mailto:bar")
+    p.DELEGATED_TO = ("mailto:foo", "mailto:bar")
+    assert p.DELEGATED_TO == ("mailto:foo", "mailto:bar")
+
 
 def test_delete_delegation_to(p):
-    p.DELEGATED_TO = ("mailto:foo","mailto:bar")
+    p.DELEGATED_TO = ("mailto:foo", "mailto:bar")
     del p.DELEGATED_TO
     assert p.DELEGATED_TO == ()
+
 
 @pytest.mark.parametrize(
     ("value", "expected"),
     [
         ((), ""),
         (("mailto:foo",), 'DELEGATED-TO="mailto:foo"'),
-        (("mailto:foo","mailto:bar"), 'DELEGATED-TO="mailto:foo","mailto:bar"'),
+        (("mailto:foo", "mailto:bar"), 'DELEGATED-TO="mailto:foo","mailto:bar"'),
         (('mailto:"asd"',), "DELEGATED-TO=\"mailto:^'asd^'\""),
-    ]
+    ],
 )
 def test_serialize_delegation_to(p, value, expected):
     p.DELEGATED_TO = value
@@ -115,19 +123,20 @@ def test_serialize_delegation_to(p, value, expected):
         (0, FBTYPE.BUSY_UNAVAILABLE),
         (1, FBTYPE.BUSY),
         (2, FBTYPE.FREE),
-    ]
+    ],
 )
 def test_get_fbtype(calendars, index, expected):
     fb = calendars.issue_798_freebusy.freebusy[index]
-    p : vPeriod = fb["FREEBUSY"]
+    p: vPeriod = fb["FREEBUSY"]
     if isinstance(p, list):
         p = p[0]
     assert expected == p.FBTYPE
 
 
-@pytest.fixture()
+@pytest.fixture
 def addr():
     return vCalAddress("mailto:foo")
+
 
 def test_partstat_get(addr: vCalAddress):
     """test the default partstat"""
@@ -142,6 +151,7 @@ def test_set_the_partstat(addr: vCalAddress):
 def test_this_and_future():
     assert vDatetime(datetime(2019, 12, 10)).RANGE is None
 
+
 def test_this_and_future_set():
     d = vDatetime(datetime(2019, 12, 10))
     d.RANGE = RANGE.THISANDFUTURE
@@ -151,16 +161,16 @@ def test_this_and_future_set():
 def test_rsvp_default(addr):
     assert not addr.RSVP
 
+
 @pytest.mark.parametrize("rsvp", [True, False])
 def test_set_rsvp(addr: vCalAddress, rsvp):
     addr.RSVP = rsvp
-    assert addr.RSVP == rsvp
+    assert rsvp == addr.RSVP
     assert addr.params["RSVP"] == ("TRUE" if rsvp else "FALSE")
 
 
 def test_sent_by(addr: vCalAddress):
     assert addr.SENT_BY is None
-
 
 
 def test_set_sent_by(addr: vCalAddress):
@@ -171,7 +181,7 @@ def test_set_sent_by(addr: vCalAddress):
 
 
 @pytest.mark.parametrize("tzid", [None, "Europe/Berlin"])
-def test_tzid(tzid, tzp:TZP):
+def test_tzid(tzid, tzp: TZP):
     dt = vDatetime(tzp.localize(datetime(2019, 12, 10), tzid))
     assert dt.TZID is None
 
@@ -181,13 +191,13 @@ def test_tzid(tzid, tzp:TZP):
     [
         (0, RELTYPE.PARENT),
         (1, RELTYPE.SIBLING),
-    ]
+    ],
 )
 def test_reltype_example(calendars, index, reltype):
     """The reltype parameter in the examples."""
     event = calendars.issue_798_related_to.events[index]
     print(event.to_ical().decode())
-    r : vText = event["RELATED-TO"]
+    r: vText = event["RELATED-TO"]
     print(r)
     print(r.params)
     assert reltype == r.RELTYPE
