+++ src/icalendar/tests/test_issue_802.py
@@ -7,17 +7,21 @@ https://github.com/collective/icalendar/issues/802
 
 import pytest
 
-from icalendar import Component, Event, Journal, Todo
+from icalendar.cal.component import Component
+from icalendar.cal.event import Event
+from icalendar.cal.journal import Journal
+from icalendar.cal.todo import Todo
 
 
 @pytest.fixture(params=[0, None])
 def default_sequence(request):
     return request.param
 
+
 @pytest.fixture(params=[Event, Journal, Todo])
 def component(request, default_sequence) -> Component:
     """Return a component."""
-    component : Component = request.param()
+    component: Component = request.param()
     if default_sequence is not None:
         component["SEQUENCE"] = default_sequence
     return component
@@ -46,6 +50,7 @@ def test_delete_sequence_default(component: Component):
     """Delete the value."""
     del component.sequence
     assert component.sequence == 0
+
 
 def test_delete_sequence_with_value(component: Component):
     """Delete the value."""
