+++ src/icalendar/tests/test_rfc_7986_categories.py
@@ -11,9 +11,13 @@ from typing import Union
 
 import pytest
 
-from icalendar import Calendar, Event, Journal, Todo
+from icalendar.cal.calendar import Calendar
+from icalendar.cal.event import Event
+from icalendar.cal.journal import Journal
+from icalendar.cal.todo import Todo
 
 CTJE = Union[Calendar, Event, Journal, Todo]
+
 
 @pytest.fixture(params=[Event, Calendar, Todo, Journal])
 def component(request):
@@ -32,10 +36,12 @@ def test_add_one_category(component: CTJE):
     component.add("categories", "Lecture")
     assert component.categories == ["Lecture"]
 
+
 def test_add_multiple_categories(component: CTJE):
     """Add categories."""
     component.add("categories", ["Lecture", "Workshop"])
     assert component.categories == ["Lecture", "Workshop"]
+
 
 def test_set_categories(component: CTJE):
     """Set categories."""
