+++ src/icalendar/tests/test_timezoned.py
@@ -3,6 +3,11 @@ import datetime
 import dateutil.parser
 
 import icalendar
+import icalendar.cal.calendar
+import icalendar.cal.event
+import icalendar.cal.timezone
+import icalendar.cal.TimezoneDaylight
+import icalendar.cal.TimezoneStandard
 from icalendar.prop import tzid_from_dt
 
 
@@ -31,7 +36,7 @@ def test_create_from_ical(calendars, other_tzp):
 
 
 def test_create_to_ical(tzp):
-    cal = icalendar.Calendar()
+    cal = icalendar.cal.calendar.Calendar()
 
     cal.add("prodid", "-//Plone.org//NONSGML plone.app.event//EN")
     cal.add("version", "2.0")
@@ -40,18 +45,18 @@ def test_create_to_ical(tzp):
     cal.add("x-wr-relcalid", "12345")
     cal.add("x-wr-timezone", "Europe/Vienna")
 
-    tzc = icalendar.Timezone()
+    tzc = icalendar.cal.timezone.Timezone()
     tzc.add("tzid", "Europe/Vienna")
     tzc.add("x-lic-location", "Europe/Vienna")
 
-    tzs = icalendar.TimezoneStandard()
+    tzs = icalendar.cal.TimezoneStandard.TimezoneStandard()
     tzs.add("tzname", "CET")
     tzs.add("dtstart", datetime.datetime(1970, 10, 25, 3, 0, 0))
     tzs.add("rrule", {"freq": "yearly", "bymonth": 10, "byday": "-1su"})
     tzs.add("TZOFFSETFROM", datetime.timedelta(hours=2))
     tzs.add("TZOFFSETTO", datetime.timedelta(hours=1))
 
-    tzd = icalendar.TimezoneDaylight()
+    tzd = icalendar.cal.TimezoneDaylight.TimezoneDaylight()
     tzd.add("tzname", "CEST")
     tzd.add("dtstart", datetime.datetime(1970, 3, 29, 2, 0, 0))
     tzs.add("rrule", {"freq": "yearly", "bymonth": 3, "byday": "-1su"})
@@ -62,7 +67,7 @@ def test_create_to_ical(tzp):
     tzc.add_component(tzd)
     cal.add_component(tzc)
 
-    event = icalendar.Event()
+    event = icalendar.cal.event.Event()
     event.add(
         "dtstart",
         tzp.localize(datetime.datetime(2012, 2, 13, 10, 00, 00), "Europe/Vienna"),
@@ -138,11 +143,13 @@ def test_create_america_new_york(calendars, tzp):
     dt = cal.events[0].start
     assert tzid_from_dt(dt) in ("custom_America/New_York", "EDT")
 
+
 def test_create_america_new_york_forward_reference(calendars, tzp):
     """testing America/New_York variant with VTIMEZONE as a forward reference"""
     cal = calendars.america_new_york_forward_reference
-    dt = cal.walk('VEVENT')[0]['DTSTART'][0].dt
-    assert tzid_from_dt(dt) in ('custom_America/New_York_Forward_reference', "EDT")
+    dt = cal.walk("VEVENT")[0]["DTSTART"][0].dt
+    assert tzid_from_dt(dt) in ("custom_America/New_York_Forward_reference", "EDT")
+
 
 def test_america_new_york_with_pytz(calendars, tzp, pytz_only):
     """Create a custom timezone with pytz and test the transition times."""
@@ -170,12 +177,12 @@ def test_america_new_york_with_pytz(calendars, tzp, pytz_only):
         datetime.timedelta(-1, 72000),
         datetime.timedelta(0, 3600),
         "EDT",
-    ) in tz._tzinfos.keys()
+    ) in tz._tzinfos
     assert (
         datetime.timedelta(-1, 68400),
         datetime.timedelta(0),
         "EST",
-    ) in tz._tzinfos.keys()
+    ) in tz._tzinfos
 
 
 fiji_transition_times = [
@@ -315,12 +322,12 @@ def test_create_pacific_fiji(calendars, pytz_only):
         datetime.timedelta(0, 46800),
         datetime.timedelta(0, 3600),
         "custom_Pacific/Fiji_19981101T020000_+1200_+1300",
-    ) in tz._tzinfos.keys()
+    ) in tz._tzinfos
     assert (
         datetime.timedelta(0, 43200),
         datetime.timedelta(0),
         "custom_Pacific/Fiji_19990228T030000_+1300_+1200",
-    ) in tz._tzinfos.keys()
+    ) in tz._tzinfos
 
 
 # these are the expected offsets before and after the fiji_transition_times
@@ -396,7 +403,7 @@ def test_transition_times_fiji(tzp, timezones):
     """The transition times are computed."""
     tz = timezones.pacific_fiji.to_tz(tzp, lookup_tzid=False)
     offsets = []  # [(before, after), ...]
-    for i, transition_time in enumerate(fiji_transition_times):
+    for _i, transition_time in enumerate(fiji_transition_times):
         before_after_offset = []
         for offset in (datetime.timedelta(hours=-1), datetime.timedelta(hours=+1)):
             time_in_timezone = tzp.localize(transition_time + offset, tz)
