+++ src/icalendar/timezone/zoneinfo.py
@@ -22,7 +22,8 @@ from dateutil.tz.tz import _tzicalvtz
 from .provider import TZProvider
 
 if TYPE_CHECKING:
-    from icalendar import cal, prop
+    from icalendar import prop
+    from icalendar.cal import Timezone
     from icalendar.prop import vDDDTypes
 
 
@@ -63,7 +64,7 @@ class ZONEINFO(TZProvider):
             # zoninfo does not know any transition dates after 2038
             rrule._until = datetime(2038, 12, 31, tzinfo=self.utc)
 
-    def create_timezone(self, tz: cal.Timezone) -> tzinfo:
+    def create_timezone(self, tz: Timezone.Timezone) -> tzinfo:
         """Create a timezone from the given information."""
         try:
             return self._create_timezone(tz)
@@ -76,13 +77,13 @@ class ZONEINFO(TZProvider):
                     if attr.lower().startswith("x-"):
                         sub.pop(attr)
             for sub in tz.subcomponents:
-                start : vDDDTypes = sub.get("DTSTART")
+                start: vDDDTypes = sub.get("DTSTART")
                 if start and is_date(start.dt):
                     # ValueError: Unsupported DTSTART param in VTIMEZONE: VALUE=DATE
                     sub.DTSTART = to_datetime(start.dt)
             return self._create_timezone(tz)
 
-    def _create_timezone(self, tz: cal.Timezone) -> tzinfo:
+    def _create_timezone(self, tz: Timezone.Timezone) -> tzinfo:
         """Create a timezone and maybe fail"""
         file = StringIO(tz.to_ical().decode("UTF-8", "replace"))
         return tzical(file).get()
@@ -116,7 +117,7 @@ def pickle_rrule_with_cache(self: rrule):
         "freq": self._freq,
         "until": self._until,
         "wkst": self._wkst,
-        "cache": False if self._cache is None else True,
+        "cache": self._cache is not None,
     }
     new_kwargs.update(self._original_rule)
     # from https://stackoverflow.com/a/64915638/1320237
@@ -137,7 +138,7 @@ def pickle_rruleset_with_cache(rs: rruleset):
         rs._rdate,
         rs._exrule,
         rs._exdate,
-        False if rs._cache is None else True,
+        rs._cache is not None,
     )
